fix: Implement ILaunchActivatedEventArgs and share the activation parser
LaunchActivatedEventArgs already had Arguments and TileId but did not declare
the live ILaunchActivatedEventArgs, so the cast Microsoft's own AppLifecycle
sample uses to read a Launch payload returned null. ProtocolActivatedEventArgs
already implemented its counterpart.

Moves TryParseActivationUri out of the .wasm.cs file: only registerProtocolHandler
is platform-bound, and keeping the parser WebAssembly-only meant its tests had to
reach it by reflection and could never execute. Drops the overload the host stopped
calling.
